% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pre_process.R
\name{data_pre_process}
\alias{data_pre_process}
\title{Data preprocessing on the timeseries data}
\usage{
data_pre_process(
  timeseries,
  adm_level = "adm0",
  date_from = min(timeseries$date)
)
}
\arguments{
\item{timeseries}{cleaned timeseries from get_timeseries}

\item{adm_level}{sets the adm level for which data to pull with options
adm0, adm1, adm2}

\item{date_from}{date from which to start the analysis}
}
\value{
tibble nested by specified adm_level
}
\description{
Data preprocessing on the timeseries data
}
\details{
Wrangle the data based on what adm level the analysis is performed.
Data will be nested by adm level to perform multi-country or multi-region
modelling across a nested tibble. This function will not be needed if only
ad hoc looks at individual countries are required.
}
