% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gen_time.R
\name{get_gen_time}
\alias{get_gen_time}
\title{Set generation time distribution}
\usage{
get_gen_time(mean = 5, sd = 7.071068, dist = "gamma", max_value = 10)
}
\arguments{
\item{mean}{mean of the generation time distribution. If specified in a
scale and shape format, the gammaParamsConvert {ConnMatTools} function can be used
to convert to mean}

\item{sd}{standard deviation of the generation time distribution. If specified in a
scale and shape format, the gammaParamsConvert {ConnMatTools} function can be used
to convert to sd}

\item{max_value}{specifies the maximum possible value. Defaults to 10 days.}

\item{distribution}{the probability distribution used. Defaults to Gamma.}
}
\value{
list including the mean, sd and distibution type for the generation time
}
\description{
Set generation time distribution
}
\details{
This function will return information on a generation time
distribution for use with the epinow package. A default distribution is
set for dengue from the literature. The mean, sd and distribution type
can be specified manually
}
