% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inc_period.R
\name{get_inc_period}
\alias{get_inc_period}
\title{Set incubation period distribution}
\usage{
get_inc_period(mean = 1.4, sd = 0.1, dist = "gamma", max_value = 6)
}
\arguments{
\item{mean}{mean of the incubation period distribution. If specified in a
scale and shape format, the gammaParamsConvert {ConnMatTools} function can be used
to convert to mean}

\item{sd}{standard deviation of the incubation period distribution. If specified in a
scale and shape format, the gammaParamsConvert {ConnMatTools} function can be used
to convert to sd}

\item{distribution}{the probability distribution used. Defaults to Gamma.}
}
\value{
list including the mean, sd and distribution type for the incubation period
}
\description{
Set incubation period distribution
}
\details{
This function will return information on an incubation period
distribution for use with the epinow package.  A default distribution is
set for dengue from the literature. The mean, sd and distribution type
can be specified manually.
}
\author{
Martina McMenamin
}
