% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nowcast.R
\name{get_nowcast}
\alias{get_nowcast}
\title{Perform estimation and create and store figures for nowcast and/or forecast}
\usage{
get_nowcast(
  data_rep,
  adm_names,
  generation_time,
  incubation_period,
  reporting_delay,
  horizon,
  nowcast = FALSE,
  rep_frequency = "weekly",
  adm_level = adm_level,
  output_dir = here("outputs"),
  date_from = min(data_rep$date),
  store_model_res = FALSE,
  create_report = TRUE
)
}
\arguments{
\item{data_rep}{nested data frame with the observed data for the country or region}

\item{generation_time}{mean and sd of the generation time distribution}

\item{incubation_period}{mean and sd of the incubation period distribution}

\item{reporting_delay}{mean and sd of the reporting delay distribution. If NULL then
nowcast is not performed and only forecast is returned}

\item{horizon}{number of days to forecast}

\item{adm_level}{the adm level for which the analysis is performed}

\item{output_dir}{the directory to store the figures and results}

\item{date_from}{the start date from which to fit the model}

\item{store_model_res}{if TRUE stores the model results as an rds file. Defaults to FALSE}
}
\description{
Perform estimation and create and store figures for nowcast and/or forecast
}
\details{
Uses observed data and specified delay distributions to estimate
the nowcast and/or forecast for a given country or region. Figures are created
and stored in a specified directory as well as model outputs if set by the user.
}
