% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{get_timeseries}
\alias{get_timeseries}
\title{Get clean timeseries of cases for estimation at a regular reporting interval
and reformat for epinow use}
\usage{
get_timeseries(data, date_var = "date", case_var = "cases", adm_level = "adm0")
}
\arguments{
\item{adm_level}{sets the adm level for which data to pull with options adm0, adm1, adm2}

\item{file_path}{specifies the file path from which to read the cleaned timeseries}
}
\value{
cleaned timeseries dataset for all countries
}
\description{
Get clean timeseries of cases for estimation at a regular reporting interval
and reformat for epinow use
}
\details{
This function will pull the timeseries from the data folder and reformat
for epinow use - this function could be tailored for a given country's dataset format
}
