% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_forecast.R
\name{run_forecast}
\alias{run_forecast}
\title{Run short term forecast using the epinow2 package for estimation}
\usage{
run_forecast(
  data,
  adm_level = "adm0",
  rep_freq = "weekly",
  nowcast = FALSE,
  generation_time = get_gen_time(),
  incubation_period = get_inc_period(),
  reporting_delay = get_rep_delay(),
  forecast = TRUE,
  horizon = 7,
  create_report = TRUE,
  date_from = min(data$date),
  output_dir = here("outputs"),
  store_model_res = FALSE,
  date_var = "date",
  case_var = "cases"
)
}
\arguments{
\item{data}{data frame with the observed data for the country or region}
}
\value{
figures for daily or weekly analysis and a report in Word format
}
\description{
Creates figures and a report on nowcast and/or short-term forecast for a given
country. This uses Epinow2 package for estimation but allows more flexibility
in application relevant to who uses such as weekly data, report creation for policy
makers and the option to include only a forecast with no nowcast when reporting
delays are not expected or counts are derived from cumulative data with no back-corrections.
}
\seealso{
[Epinow2::epinow()] which this function wraps
}
