% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_reported_week.R
\name{viz_reported_week}
\alias{viz_reported_week}
\title{Visualise the reported and estimated reported cases by date of report}
\usage{
viz_reported_week(now_estimates, adm_names, reporting_freq)
}
\arguments{
\item{now_estimates}{List of estimates from the nowcasting/forecasting}

\item{adm_level}{sets the adm level for which data to pull with options adm0, adm1}
}
\description{
Visualise the reported and estimated reported cases by date of report
}
\details{
A function to visualise estimated reported weekly cases as epinow
only shows as a continuous timeseries with no weekly periodicity.
This function aggregates reported case estimates to the weekly level and
overlays this on the weekly reported case epicurve.This is also needed as
epinow show breaks in estimates which we don't want across short time frames
}
