% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_R.R
\name{viz_Rt}
\alias{viz_Rt}
\title{Vizualise the cases by infection date}
\usage{
viz_Rt(now_estimates, adm_level)
}
\arguments{
\item{now_estimates}{List of estimates from the nowcasting}

\item{results}{Identifies the figure to generate. Options are 'infections', 'Rt',
'growth_rate' or 'reported.}
}
\value{
Figure showing reported weekly cases epi curve with estimated daily
cases by infection date overlaid
}
\description{
Vizualise the cases by infection date
}
\details{
This function will vizualise the nowcast estimates for daily cases by
infection date by extracting this from the preexisting epinow format
}
\author{
Martina McMenamin
}
