% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rep_delay.R
\name{get_rep_delay}
\alias{get_rep_delay}
\title{Set reporting delay distribution}
\usage{
get_rep_delay(mean = 2, sd = 1, dist = "lognormal", max_value = 10)
}
\arguments{
\item{mean}{mean of the reporting delay distribution. If a lognormal
distribution is specified, the convert_to_logmean {EpiNow2} function can be used
to convert to format needed}

\item{sd}{standard deviation of the reporting delay distribution. If a lognormal
distribution is specified, the convert_to_logsd {EpiNow2} function can be used
to convert to format needed}

\item{distribution}{the probability distribution used. Defaults to lognormal.}
}
\value{
list including the mean, sd and distribution type for the reporting delay distribution
}
\description{
Set reporting delay distribution
}
\details{
This distribution is used when a nowcast is needed to correct for
reporting delays in the data. Ideally this should be estimated for a given country
or region using timestamped datasets in order to properly account for the reporting
delays. Otherwise an unknown reporting delay distribution can be assumed but
results will not be as accurate.
}
